+++ static/js/local.js
@@ -2864,7 +2864,8 @@ WorkerBridgeServer.prototype.onWorkerLog = function(message) {
 
 	// Gets an access token from the provider & user using a popup
 	// - Best if called within a DOM click handler, as that will avoid popup-blocking
-	Relay.prototype.requestAccessToken = function() {
+	// - `opts.guestof`: optional string, the host userid providing the guest account. If specified, attempts to get a guest session
+	Relay.prototype.requestAccessToken = function(opts) {
 		// Start listening for messages from the popup
 		if (!this.messageFromAuthPopupHandler) {
 			this.messageFromAuthPopupHandler = (function(e) {
@@ -2897,7 +2898,9 @@ WorkerBridgeServer.prototype.onWorkerLog = function(message) {
 
 		// Open interface in a popup
 		// :HACK: because popup blocking can only be avoided by a syncronous popup call, we have to manually construct the url (it burns us)
-		window.open(this.getProvider() + '/session/' + this.config.app);
+		var url = this.getProvider() + '/session/' + this.config.app;
+		if (opts && opts.guestof) { url += '?guestof='+encodeURIComponent(opts.guestof); }
+		window.open(url);
 	};
 
 	// Fetches users from p2pw service